requirements.txt
pytest>=5.0.0
allure-pytest>=2.8.18
selene>=2.0.0b4